backend/requirements.txt
# ──────────────────────────────────────
# Core Framework
# ──────────────────────────────────────
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-dotenv==1.0.1
pydantic==2.10.4
pydantic-settings==2.7.1

# ──────────────────────────────────────
# Database
# ──────────────────────────────────────
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.1
pgvector==0.3.6

# ──────────────────────────────────────
# Authentication
# ──────────────────────────────────────
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1

# ──────────────────────────────────────
# AI / LLM
# ──────────────────────────────────────
openai==1.58.1
crewai==0.95.0
crewai-tools==0.17.0

# ──────────────────────────────────────
# Redis
# ──────────────────────────────────────
redis[hiredis]==5.2.1

# ──────────────────────────────────────
# File Processing
# ──────────────────────────────────────
PyMuPDF==1.25.1
python-multipart==0.0.20

# ──────────────────────────────────────
# Email
# ──────────────────────────────────────
resend==2.7.0

# ──────────────────────────────────────
# Cloud Storage
# ──────────────────────────────────────
boto3==1.36.2

# ──────────────────────────────────────
# Testing
# ──────────────────────────────────────
pytest==8.3.4
pytest-asyncio==0.25.0
httpx==0.28.1

# ──────────────────────────────────────
# Dev Tools
# ──────────────────────────────────────
ruff==0.9.1
mypy==1.14.1
